requirements.txt
#torch>=1.1.0
#torchvision>=0.3.0,<=0.11.2
segmentation-models-pytorch==0.3.1
albumentations==1.3.0
shapely>=1.8.5
aeronet==0.0.18
scikit-learn==0.24.2
pandas==1.1.5
ttach==0.0.3
addict==2.4.0
fire==0.4.0
# Pillow>=6.1.0
rasterio<=1.2.5
boto3==1.23.10

pycksum=0.3.3
wandb


##tensorflow==2.7.0  # for tensorboard
##tensorboard>=1.11.0
#keras_preprocessing==1.1.2
#keras_applications==1.0.8
#gast==0.4.0
#opt_einsum
#astunparse
#flatbuffers
#wrapt